Create Design Matrices

Create design matrices for fixed effects and temporal, age, spatial and random effects, for both medical and traditional circumcision.

create_design_matrices(out, k_dt = 5)

Arguments

out

Shell dataset (outputted by create_shell_dataset with a row for every unique record in circumcision survey data for a given area. Also includes empirical estimates for circumcision estimates for each unique record.

k_dt

Age knot spacing in spline definitions, Default: 5

Value

List of design matrices for fixed and random effects for medical and traditional circumcision.

See also