Create Matrix to Estimate Cumulative Hazard Rate

Create a matrix to estimate the cumulative hazard rate needed for survival analysis by age and time. The option to include an additional stratification variable is also available, creating a 3D hazard function.

create_integration_matrix_agetime(
  dat,
  subset = NULL,
  time1 = "time1",
  time2 = "time2",
  timecaps = c(1, Inf),
  Ntime = NULL,
  age = "age",
  Nage = NULL,
  strat = NULL,
  Nstrat = NULL
)

Arguments

dat

Dataset used for modelling.

subset

Subset for dataset, Default: NULL

time1

Variable name for time of birth, Default: "time1"

time2

Variable name for time circumcised or censored, Default: "time2"

timecaps

Window to fix temporal dimension before and after, Default: c(1, Inf)

Ntime

Number of time points (if NULL, function will calculate), Default: NULL

age

Variable with age circumcised or censored, Default: "age"

Nage

Number of age groups (if NULL, function will calculate), Default: NULL

strat

Variable to stratify by in using a 3D hazard function, Default: NULL

Nstrat

Number of stratification groups (if NULL, function will calculate), Default: NULL

Value

Matrix for selecting instantaneous hazard rate.

See also