Minimise Fit Object Size

Return minimised fit object. Often useful when saving the fit object for later aggregation.

minimise_fit_obj(fit, dat_tmb, parameters)

Arguments

fit

Optional "small" fit object with no sample. Specifying fit means you do not need to specify dat_tmb or parameters, as argument specifications will be overridden by those stored in fit.

dat_tmb

list of data required for model fitting, outputted by threemc_prepare_model_data, which includes:

design_matricesIncludesX_fixed_mmc, X_fixed_tmc, X_time_mmc, X_age_mmc, X_age_tmc, X_space_mmc, X_space_tmc, X_agetime_mmc, X_agespace_mmc, X_agespace_tmc, X_spacetime_mmc. Design Create design matrices for fixed effects and temporal, age, space and interaction random effects

integration matricesIncludes IntMat1, IntMat2. Integration matrices for selecting the instantaneous hazard rate.

survival matricesIncludes A_mmc, A_tmc, A_mc, B, C. Survival matrices for MMC, TMC, censored and left censored

Q_spacePrecision/Adjacency matrix for the spatial random effects.

parameters

list of fixed and random model parameters.

Value

Object of class "naomi_fit".