Define circumcision type

Using circ_who and circ_where, determines survey type.

Using circ_who and circ_where, determines survey type.

Join surveys with areas for each level leading up to the desired area level. This distributes granular surveys to desired area_level using parent_ids.

compare_predictions(actual, predictions)

find_circ_type(survey_circumcision)

reassign_survey_level(
  survey_circumcision,
  areas,
  area_lev,
  remove_add_cols = TRUE
)

Arguments

actual

vector of observed values.

predictions

dataframe/tibble/data.table or matrix with a row for each value in actual, and a column for each prediction.

survey_circumcision

Information on male circumcision status from surveys. If this is a list or contains more than one country, the function is performed for each country present, returning a list.

areas

sf shapefiles for specific country/region.

area_lev

Desired admin boundary level to perform the analysis on.

remove_add_cols

Only keep columns originally present in survey_circumcision, Default = TRUE.

Value

List of:

"summary_stats_df"pointwise prediction stats as a data.frame

"summary_stats"mean/summed overall prediction stats

Surveys with circumcision type labelled by the column type.

Surveys with records reassigned to area_lev.