Use model shell dataset to estimate empirical circumcision rates

Takes the shell dataset with a row for every unique area ID, area name, year and circumcision age in survey data outputed by create_shell_dataset and returns the empirical circumcision rates for each row, aggregated to age groups from single ages. Also converts from wide format to long format.

threemc_empirical_rates(
  out,
  areas,
  area_lev,
  populations,
  age_groups = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "54-59", "0+", "10+", "15+", "15-24", "10-24", "15-29",
    "10-29", "15-39", "10-39", "15-49", "10-49")
)

Arguments

out

Shell dataset outputted by create_shell_dataset

areas

sf shapefiles for specific country/region.

area_lev

Desired admin boundary level to perform the analysis on.

populations

data.frame containing populations for each region in tmb fits.

age_groups

Age groups to aggregate by, Default: c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "54-59", "15-24", "10-24", "15-29", "10-29", "15-39", "10-39", "15-49", "10-49" )

See also