Produce Data Matrices for Modelling

Create data for modelling. Output detailed below.

threemc_prepare_model_data(
  out,
  areas,
  area_lev = NULL,
  aggregated = TRUE,
  weight = "population",
  k_dt = 5,
  paed_age_cutoff = NULL,
  rw_order = NULL,
  ...
)

Arguments

out

Shell dataset (outputted by create_shell_dataset with a row for every unique record in circumcision survey data for a given area. Also includes empirical estimates for circumcision estimates for each unique record.

areas

sf shapefiles for specific country/region.

area_lev

PSNU area level for specific country.

aggregated

agggregated = FALSE treats every area_id as its own object, allowing for the use of surveys for lower area hierarchies. aggregated = TRUE means we only look at area level of interest.

weight

variable to weigh circumcisions by when aggregating for lower area hierarchies (only applicable for aggregated = TRUE)

k_dt

Age knot spacing in spline definitions, Default: 5

paed_age_cutoff

Age at which to split MMC design matrices between paediatric and non-paediatric populations, the former of which are constant over time. Set to NULL if not desired, Default: NULL

rw_order

Order of the random walk used for temporal precision matrix. Setting to NULL assumes you wish to specify an AR 1 temporal prior. Default: NULL

...

Additional arguments to be passed to functions which create matrices.

Value

list of data required for model fitting, including:

design_matricesIncludesX_fixed_mmc, X_fixed_tmc, X_time_mmc, X_age_mmc, X_age_tmc, X_space_mmc, X_space_tmc, X_agetime_mmc, X_agespace_mmc, X_agespace_tmc, X_spacetime_mmc. Design Create design matrices for fixed effects and temporal, age, space and interaction random effects

integration matricesIncludes IntMat1, IntMat2. Integration matrices for selecting the instantaneous hazard rate.

survival matricesIncludes A_mmc, A_tmc, A_mc, B, C. Survival matrices for MMC, TMC, censored and left censored

Q_spacePrecision/Adjacency matrix for the spatial random effects.